refactor: new onboarding workflow
